--- packages/peregrine/src/Peregrine/index.js
@@ -1,1 +1,0 @@
-export { default } from './peregrine';
