+++ packages/peregrine/src/PeregrineContextProvider/index.js
@@ -1,0 +1,1 @@
+export { default } from './peregrineContextProvider';
