+++ packages/peregrine/src/index.js
@@ -21,7 +21,9 @@ export { default as List, Items, Item } from './List';
 export { default as Page } from './Page';
 export { default as Price } from './Price';
 export { default as Router } from './Router';
-export { default as Peregrine } from './Peregrine';
+export {
+    default as PeregrineContextProvider
+} from './PeregrineContextProvider';
 
 // misc
 export { RestApi };
