+++ packages/venia-concept/src/components/App/contextProvider.js
@@ -1,0 +1,26 @@
+import React from 'react';
+
+import {
+    PeregrineContextProvider as Peregrine,
+    ToastContextProvider,
+    WindowSizeContextProvider
+} from '@magento/peregrine';
+
+/**
+ * List of context providers that are required to run Venia
+ *
+ * @property {React.Component[]} contextProviders
+ */
+const contextProviders = [
+    Peregrine,
+    WindowSizeContextProvider,
+    ToastContextProvider
+];
+
+const AppContextProvider = ({ children }) => {
+    return contextProviders.reduceRight((memo, ContextProvider) => {
+        return <ContextProvider>{memo}</ContextProvider>;
+    }, children);
+};
+
+export default AppContextProvider;
