+++ packages/venia-concept/src/components/App/index.js
@@ -1,1 +1,2 @@
 export { default } from './container';
+export { default as AppContextProvider } from './contextProvider';
