+++ packages/venia-concept/src/index.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { setContext } from 'apollo-link-context';
-import { Peregrine, Util } from '@magento/peregrine';
+import { Util } from '@magento/peregrine';
 import { Adapter } from 'src/drivers';
 import store from 'src/store';
 import app from 'src/actions/app';
-import App from 'src/components/App';
+import App, { AppContextProvider } from 'src/components/App';
 import './index.css';
 
 const { BrowserPersistence } = Util;
@@ -36,9 +36,9 @@ ReactDOM.render(
         apollo={{ link: authLink.concat(Adapter.apolloLink(apiBase)) }}
         store={store}
     >
-        <Peregrine>
+        <AppContextProvider>
             <App />
-        </Peregrine>
+        </AppContextProvider>
     </Adapter>,
     document.getElementById('root')
 );
